config refactoring
