+++ commands/controller/dbrestore.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const config = require('../../config.json')
+const config = require('../../config.js')
 const path = require('path')
 const mongoose = require('mongoose')
 const storage = require('../../util/storage.js')
