+++ commands/controller/setgame.js
@@ -1,4 +1,4 @@
-const config = require('../../config.json')
+const config = require('../../config.js')
 const log = require('../../util/logger.js')
 
 function getGame (message) {
