+++ commands/rssdate.js
@@ -1,4 +1,4 @@
-const config = require('../config.json')
+const config = require('../config.js')
 const moment = require('moment-timezone')
 const storage = require('../util/storage.js')
 const currentGuilds = storage.currentGuilds
