+++ commands/rsslist.js
@@ -1,4 +1,4 @@
-const config = require('../config.json')
+const config = require('../config.js')
 const storage = require('../util/storage.js')
 const log = require('../util/logger.js')
 const currentGuilds = storage.currentGuilds
