+++ commands/rssroles.js
@@ -1,7 +1,7 @@
 const Discord = require('discord.js')
 const dbOps = require('../util/dbOps.js')
 const filters = require('./util/filters.js')
-const config = require('../config.json')
+const config = require('../config.js')
 const log = require('../util/logger.js')
 const currentGuilds = require('../util/storage.js').currentGuilds
 const MenuUtils = require('../structs/MenuUtils.js')
