+++ commands/util/filters.js
@@ -1,5 +1,5 @@
 const dbOps = require('../../util/dbOps.js')
-const config = require('../../config.json')
+const config = require('../../config.js')
 const filterTypes = [
   { show: 'Title', use: 'title' },
   { show: 'Description', use: 'description' },
@@ -56,7 +56,7 @@ async function inputFilterFn (m, data) {
     } else invalidItems += `\n${item}`
   })
 
-  
+
   if (!role) {
     log.command.info(`New filter(s) [${addedList.trim().split('\n')}] being added to '${chosenFilterType}' for ${source.link}`, m.guild)
     await dbOps.guildRss.update(guildRss, true)
