+++ rss/initialize.js
@@ -1,7 +1,7 @@
 const requestStream = require('./request.js')
 const FeedParser = require('feedparser')
 const dbOps = require('../util/dbOps.js')
-const config = require('../config.json')
+const config = require('../config.js')
 const dbCmds = require('./db/commands.js')
 const storage = require('../util/storage.js')
 const currentGuilds = storage.currentGuilds
