+++ rss/request.js
@@ -1,6 +1,6 @@
 const fetch = require('node-fetch')
 const cloudscraper = require('cloudscraper') // For cloudflare
-const config = require('../config.json')
+const config = require('../config.js')
 
 module.exports = async (link, cookies) => {
   const options = {
